feat: restructure and cleanup
