+++ packages/store/src/collection/index.spec.ts
@@ -7,8 +7,8 @@ import sinon from 'sinon'
 import mingo from 'mingo'
 import { collect } from 'streaming-iterables'
 import { encode } from 'cbor-sync'
-import { Collection, ReadBatch, WriteBatch, CollectionKey } from './collection'
 import { FilterQuery } from './query'
+import { Collection, ReadBatch, WriteBatch, CollectionKey } from '.'
 import { ActionHandler } from '.'
 import { EntityID, Action } from '..'
