+++ packages/store/src/dispatcher/index.spec.ts
@@ -1,6 +1,6 @@
 import { expect } from 'chai'
 import { pack } from 'lexicographic-integer'
-import { Dispatcher } from './dispatcher'
+import { Dispatcher } from '.'
 import { Event } from '..'
 
 describe('Dispatcher', () => {
